refactor: move notepage
