+++ src/route/Router.tsx
@@ -1,26 +1,20 @@
-﻿import { Routes, Route, useParams } from "react-router-dom";
+﻿import { Routes, Route } from "react-router-dom";
 import HomePage from "../services/home/HomePage";
 import LandingPage from "../services/landing/LandingPage";
 import { useAuthStore } from "../store/useAuthStore";
 import LoginPage from "../services/auth/LoginPage";
-import NotePage from "../services/note/NotePage";
 import OAuthCallback from "../services/auth/OauthPage";
 import AppLayout from "../services/AppLayout";
-import NoteDetailPage from "../services/note/NoteDetailPage";
-import NoteEditPage from "../services/note/NoteEditPage";
+import NoteDetailPage from "../services/contents/note/NoteDetailPage";
+import NoteEditPage from "../services/contents/note/NoteEditPage";
 import MyPage from "../services/mypage/MyPage";
 import QuestionListPage from "../services/question/QuestionListPage";
 import QuestionDetailPage from "../services/question/QuestionDetailPage";
-import FilePage from "../services/file/FilePage";
+import FilePage from "../services/contents/FilePage";
 
 function Router() {
   const { user } = useAuthStore();
   const isLoggedIn = !!user;
-
-  const StudySlugPage = () => {
-    const { slug } = useParams();
-    return slug?.includes("강의필기") ? <NotePage /> : <FilePage />;
-  };
 
   return (
     <Routes>
@@ -58,7 +52,7 @@ function Router() {
         path="/study/:slug"
         element={
           <AppLayout>
-            <StudySlugPage />
+            <FilePage />
           </AppLayout>
         }
       />
