+++ src/services/contents/FileDetailPage.tsx
@@ -1,0 +1,5 @@
+﻿function FileDetailPage() {
+  return <></>;
+}
+
+export default FileDetailPage;
