--- src/services/file/FilePage.tsx
@@ -1,5 +1,0 @@
-﻿function FilePage() {
-  return <></>;
-}
-
-export default FilePage;
